Refactorizar código, separar en modulos y añadir readme
