+++ README.md
@@ -1,0 +1,12 @@
+# Asteroides C :rocket:
+
+## ¿Por qué AsteroidesC?
+En su momento fue un proyecto de primer semestre de la carrera.
+Está escrito en C y usa la librería para videojuegos [Allegro](https://github.com/liballeg/allegro5).
+
+## ¿Por qué código abierto?
+Espero que algun colega estudiante pueda usar este proyecto como ayuda para realizar el suyo, o bien
+que utilice el código para seguir desarrollandolo.
+
+## ¿Cómo compilar?
+Proximamente...
